Add files via upload
The warning message itself is not an error and can be ignored, but it is indicating that in future versions of Pandas, the numeric_only parameter will need to be explicitly specified when using the mean() function with groupby().

The script then continues to run, displaying the data DataFrame and some summary statistics using the describe() function. It then groups the data by category and computes the mean of each group, which is stored in the grouped_data variable. Finally, the script attempts to merge two DataFrames, data1 and data2, on the column_name column. However, there is no data1 DataFrame defined in the code, which results in a NameError being raised. It suggests that there might be a typo in the code and that data was meant to be used instead of data1.